+++ R/compare_performance.R
@@ -54,6 +54,13 @@
 #'   }
 #'
 #' @examples
+#' data(iris)
+#' lm1 <- lm(Sepal.Length ~ Species, data = iris)
+#' lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
+#' lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
+#' compare_performance(lm1, lm2, lm3)
+#' compare_performance(lm1, lm2, lm3, rank = TRUE)
+#'
 #' if (require("lme4")) {
 #'   m1 <- lm(mpg ~ wt + cyl, data = mtcars)
 #'   m2 <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
@@ -61,12 +68,6 @@
 #'   compare_performance(m1, m2, m3)
 #' }
 #'
-#' data(iris)
-#' lm1 <- lm(Sepal.Length ~ Species, data = iris)
-#' lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
-#' lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
-#' compare_performance(lm1, lm2, lm3)
-#' compare_performance(lm1, lm2, lm3, rank = TRUE)
 #' @importFrom insight is_model_supported all_models_equal get_response
 #' @importFrom bayestestR bayesfactor_models
 #' @inheritParams model_performance.lm
@@ -123,7 +124,7 @@ compare_performance <- function(..., metrics = "all", rank = FALSE, bayesfactor 
   }
 
   if (!is.null(LRTs)) {
-    LRTs <- LRTs[c("Model", "Type", "p")]
+    LRTs <- LRTs[c("Model", "Type", "df", "p")]
     LRTs$Model <- sapply(object_names, deparse)
     dfs <- merge(dfs, LRTs, by = c("Model", "Type"), all = TRUE, sort = FALSE)
   }